refactor: move seed endpoint as a system endpoint
